+++ examples/docs_snippets/docs_snippets/integrations/dagstermill/iris_notebook_op.py
@@ -1,0 +1,21 @@
+# isort: split
+# start
+
+from dagstermill import define_dagstermill_op, local_output_notebook_io_manager
+
+from dagster import file_relative_path, job
+
+k_means_iris = define_dagstermill_op(
+    name="k_means_iris",
+    notebook_path=file_relative_path("iris-kmeans.ipynb"),
+    output_notebook_name="iris_kmeans_output",
+)
+
+
+@job(
+    resource_defs={
+        "output_notebook_io_manager": local_output_notebook_io_manager,
+    }
+)
+def iris_classify():
+    k_means_iris()
