--- examples/docs_snippets/docs_snippets/legacy/data_science/iris_classify.yaml
@@ -1,5 +1,0 @@
-ops:
-  download_file:
-    config:
-      path: 'iris.data'
-      url: 'https://archive.ics.uci.edu/ml/machine-learning-databases/iris/iris.data'
