--- examples/docs_snippets/docs_snippets/legacy/data_science/iris_classify_dev.yaml
@@ -1,4 +1,0 @@
-ops:
-  k_means_iris:
-    inputs:
-      path: "iris.data"
